requirements-dev.txt
torch==2.2.2
torchvision==0.17.2
easyocr==1.7.1
opencv_contrib_python==4.10.0.84
opencv_python==4.5.5.62
opencv_python_headless==4.10.0.84
Pillow==10.4.0
segment_anything==1.0
openai
python-dotenv
numpy==1.26.4
pandas
flask
ultralytics